requirements-prod.txt
# Versions pinned with ~= (compatible release). Use pip-compile for a full lock file.
# Production requirements for Railway deployment
# Only includes packages actually needed by the dashboard app
# Full development dependencies are in requirements.txt

# Core framework
fastapi~=0.109.0
uvicorn[standard]~=0.27.0
pydantic~=2.5.0
pydantic-settings~=2.1.0
email-validator~=2.0.0

# HTTP clients (used by uvicorn/fastapi)
httpx~=0.26.0

# Database
sqlalchemy~=2.0.0
psycopg2-binary~=2.9.0
redis~=5.0.0

# Auth
PyJWT~=2.0.0
bcrypt~=4.0.0

# Templates & forms
jinja2~=3.1.0
python-multipart~=0.0.6
pyyaml~=6.0.0

# Rate limiting
slowapi~=0.1.9

# Logging
loguru~=0.7.0

# CLI/display (loaded via src/__init__)
rich~=13.7.0
tenacity~=8.2.0

# Environment
python-dotenv~=1.0.0

# Billing (top-level import in billing/routes.py)
stripe~=7.0.0

# LLM clients (try/except guarded but commonly used)
openai~=1.7.0
anthropic~=0.8.0

# GraphQL (try/except guarded)
strawberry-graphql[fastapi]~=0.216.0

# CSRF protection
starlette-csrf~=2.0.0

# Intelligence engine (top-level imports in src/intelligence/processor.py)
scikit-learn~=1.3.0
textblob~=0.17.0

# Optional: monitoring (lazy import)
sentry-sdk~=1.39.0

# Optional: caching (lazy import)
diskcache~=5.6.0

# Deployment
PyGithub~=2.1.0
